Fetch Gaza and West Bank Detainee number separately.
Changed the structure of the JSON file to allow detainee data to be
stored separately, and added a version field in case we want to do
"migrations" later on.

Updated the type description of `DayData` to list all fields explicitly,
as well as modifying the loading and data reading code to be less
obtuse.